@@ -606,7 +606,9 @@ class TestOptimizerIntegration:
 
         # Use same initial weights
         for (_name1, param1), (_name2, param2) in zip(
-            model1.named_parameters(), model2.named_parameters()
+            model1.named_parameters(),
+            model2.named_parameters(),
+            strict=True,
         ):
             param2.data = param1.data.clone()
 
@@ -644,7 +646,9 @@ class TestOptimizerIntegration:
         # Check that the models have diverged (different behavior)
         diverged = False
         for (_name1, param1), (_name2, param2) in zip(
-            model1.named_parameters(), model2.named_parameters()
+            model1.named_parameters(),
+            model2.named_parameters(),
+            strict=True,
         ):
             if not torch.allclose(param1.data, param2.data):
                 diverged = True
